+++ app/src/main/java/com/nasiat_muhib/classmate/domain/state/CreateCourseUIState.kt
@@ -5,7 +5,7 @@ import com.nasiat_muhib.classmate.core.Constants.SEMESTERS
 data class CreateCourseUIState (
     val courseCode: String = "",
     val courseTitle: String = "",
-    val courseCredit: Float = 0f,
+    val courseCredit: Double = 0.0,
     val courseSemester: String = SEMESTERS[0],
     val courseTeacherEmail: String = "",
